Put point indices at the end in tabulate()
This change makes it so that the array produced by
`GridFunction.tabulate()` for tensor output functions is indexed with
component index first and grid point index second, instead of the other
way around. This way, `np.ravel` call on the result of `tabulate()` will
produce a vector that can be multiplied by a matrix constructed from
blocks corresponding to function output components, which is easier to
construct and inspect.