Copy project files when building the Docker image
During `poetry install`, poetry attempts to create the `poetry.lock`
file with the list of resolved dependencies. Since the mount in RUN
instructions in Dockerfile are read-only, in the current setup this is
not possible, so the installation step fails unless the `poetry.lock`
file has already been present in the build context.

This change makes it so that the code is copied into the container
instead of mounting it, so the `poetry.lock` file can be created. Since
it only exists in the container, it needs to be manually copied to the
host after building the container.